Moved models to models folder. Created test bed for GattService
