+++ src/Plugin.BlueZ.ManualTest/GattServer.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Plugin.BlueZ.ManualTest
+{
+  /// <summary>Gatt Server - COMING SOON.</summary>
+  /// <remarks>
+  ///   References:
+  ///   - https://github.com/bluez/bluez/blob/master/test/example-gatt-server
+  /// </remarks>
+  public class GattServer
+  {
+    // 1. BlueZManager.GetAdapterAsync();
+    // 2. CreateProxy GATT Description
+    // 3. Create GATT Services
+    // 4. Create GATT Characteristics
+
+    /*
+     * 'org.bluez'
+     * 'org.bluez.GattManager1'
+     * 'org.freedesktop.DBus.ObjectManager'
+     * 'org.freedesktop.DBus.Properties'
+     * 
+     * 'org.bluez.GattService1'
+     * 'org.bluez.GattCharacteristic1'
+     * 'org.bluez.GattDescriptor1'
+     */
+  }
+}
