+++ src/Plugin.BlueZ/Constants.cs
@@ -6,7 +6,11 @@
     public const string AdapterInterface = "org.bluez.Adapter1";
     public const string DeviceInterface = "org.bluez.Device1";
     public const string GattServiceInterface = "org.bluez.GattService1";
+    public const string GattManagerInterface = "org.bluez.GattManager1";
     public const string GattCharacteristicInterface = "org.bluez.GattCharacteristic1";
+    public const string GattDescriptorInterface = "org.bluez.GattDescriptor1";
+    public const string LeAdvertisementInterface = "org.bluez.LEAdvertisement1";
+    public const string LeAdvertisingManagerInterface = "org.bluez.LEAdvertisingManager1";
   }
 
   // https://www.bluetooth.com/specifications/gatt/
