+++ src/Plugin.BlueZ/Models/AdapterProperties.cs
