+++ src/Plugin.BlueZ/Models/DeviceProperties.cs
